feat(scripts): chiflix

+++ scripts/publish.sh
@@ -8,7 +8,8 @@ fi
 DIR=$(dirname $(dirname $($READLINK -f $0)))
 if [ -z $DIR ]; then exit 1; fi;
 rm -fr $DIR/dist
-git clone --branch=gh-pages --depth=1 $DEPLOYMENT_REPO $DIR/dist
+git clone --branch=gh-pages --depth=1 git@github.com:chiflix/chiflix.github.io.git $DIR/dist
+if [ -z $DIR/dist ]; then exit 1; fi;
 rm -fr $DIR/dist/*
 cp -r $DIR/src/* $DIR/dist
 VERSION=`curl -is "https://github.com/chiflix/splayerx/releases/latest" | grep -E '^Location: ' | grep -oE "([0-9]+.)+[0-9]"`
